requirements.txt
django-environ==0.4.5
django-allauth==0.45.0
django-tailwind==2.2.0
crispy-tailwind==0.5.0
djangorestframework==3.12.4